+++ R/case.R
@@ -1,7 +1,8 @@
 #' Convert case of a string.
 #'
 #' @param string String to modify
-#' @param locale Locale to use for translations.
+#' @param locale Locale to use for translations. Defaults to "en" (English)
+#'   to ensure consistent default ordering across platforms.
 #' @examples
 #' dog <- "The quick brown dog"
 #' str_to_upper(dog)
@@ -9,23 +10,23 @@
 #' str_to_title(dog)
 #'
 #' # Locale matters!
-#' str_to_upper("i", "en") # English
+#' str_to_upper("i") # English
 #' str_to_upper("i", "tr") # Turkish
 #' @name case
 NULL
 
 #' @export
 #' @rdname case
-str_to_upper <- function(string, locale = "") {
+str_to_upper <- function(string, locale = "en") {
   stri_trans_toupper(string, locale = locale)
 }
 #' @export
 #' @rdname case
-str_to_lower <- function(string, locale = "") {
+str_to_lower <- function(string, locale = "en") {
   stri_trans_tolower(string, locale = locale)
 }
 #' @export
 #' @rdname case
-str_to_title <- function(string, locale = "") {
+str_to_title <- function(string, locale = "en") {
   stri_trans_totitle(string, opts_brkiter = stri_opts_brkiter(locale = locale))
 }