+++ R/sort.R
@@ -6,7 +6,8 @@
 #' @param na_last Where should \code{NA} go? \code{TRUE} at the end,
 #'   \code{FALSE} at the beginning, \code{NA} dropped.
 #' @param locale In which locale should the sorting occur? Defaults to
-#'   the current locale.
+#'   the English. This ensures that code behaves the same way across
+#'   platforms.
 #' @param numeric If \code{TRUE}, will sort digits numerically, instead
 #'    of as strings.
 #' @param ... Other options used to control sorting order. Passed on to
@@ -14,8 +15,8 @@
 #' @seealso \code{\link[stringi]{stri_order}} for the underlying implementation.
 #' @export
 #' @examples
-#' str_order(letters, locale = "en")
-#' str_sort(letters, locale = "en")
+#' str_order(letters)
+#' str_sort(letters)
 #'
 #' str_order(letters, locale = "haw")
 #' str_sort(letters, locale = "haw")
@@ -24,7 +25,7 @@
 #' str_sort(x)
 #' str_sort(x, numeric = TRUE)
 str_order <- function(x, decreasing = FALSE, na_last = TRUE,
-                      locale = "", numeric = FALSE, ...) {
+                      locale = "en", numeric = FALSE, ...) {
   stri_order(x, decreasing = decreasing, na_last = na_last,
     opts_collator = stri_opts_collator(locale, numeric = numeric, ...))
 }
@@ -32,7 +33,7 @@ str_order <- function(x, decreasing = FALSE, na_last = TRUE,
 #' @export
 #' @rdname str_order
 str_sort <- function(x, decreasing = FALSE, na_last = TRUE,
-                     locale = "", numeric = FALSE, ...) {
+                     locale = "en", numeric = FALSE, ...) {
   stri_sort(x, decreasing = decreasing, na_last = na_last,
     opts_collator = stri_opts_collator(locale, numeric = numeric, ...))
 }